build: when automatically pulling in changes from `@octokit/openapi`, automatically fix linting errors on the output
This updates the "update" task which pulls in changes to
`@octokit/openapi` to run Prettier and automatically fix code
style and formatting issues.

If we assume that the code currently in `main` is valid from a
linting perspective - which it should be, given our required
checks, and this will only affect code that was actually generated
by the update flow.