build: download OpenAPI specs direct from GitHub, not Unpkg
We've been unable to update the OpenAPI specs today and regenerate
the types because Unpkg is [returning][1] a `420` error when we
try to download the generated specs in `octokit/openapi`.

This switches to downloading them from GitHub instead using the
Release Assets API, which seems to be working better and avoids
going to a third party to get something that lives in GitHub.

[1]: https://github.com/octokit/openapi-types.ts/runs/7476119832?check_suite_focus=true